daism_dnn/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
argh=0.26.2=py37_1001 
anndata=0.6.22=py_0
scanpy=1.4.3=py_0
sklearn=0.0=pypi_0
scikit-learn=0.21.2=pypi_0
scipy=1.3.0=pypi_0
python=3.7.3=h0371630_0
numpy=1.16.3=py37h7e9f1db_0
pytorch=1.0.1=py3.7_cuda8.0.61_cudnn7.1.2_2
pandas=0.25.1=pypi_0

